pyproject.toml
[project]
name = "streambank-demo"
version = "0.1.0"
description = "StreamBank demo"
authors = [
    {name = "Oleg",email = "potrebchuk0@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "gunicorn (>=20.0,<21.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "celery-prometheus (>=1.1.0,<2.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "libpass (>=1.9.1.post0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "django-redis (>=5.4.0,<6.0.0)",
    "types-python-jose (>=3.5.0.20250531,<4.0.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "types-psycopg2 (>=2.9.21.20250516,<3.0.0.0)",
    "django-stubs (>=5.2.0,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.0"
django-stubs = "^5.2.0"
mkdocs-material = "^9.6.14"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
pymdown-extensions = "^10.15"

[tool.black]
line-length = 79
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
python_version = "3.12"
mypy_path = "backend"
plugins = "mypy_django_plugin.main"
ignore_missing_imports = true
explicit_package_bases = true

[tool.django-stubs]
django_settings_module = "config.settings"
